1724: check to see if file exists already, and don't copy it over if it does...
git-svn-id: https://svn.code.sf.net/p/autoplot/code/autoplot/trunk@20083 0c594bb0-8543-0410-a25e-d8ec440261db

+++ JythonSupport/src/org/virbo/jythonsupport/JythonUtil.java
@@ -4,6 +4,7 @@ package org.virbo.jythonsupport;
 import java.io.BufferedReader;
 import java.io.ByteArrayInputStream;
 import java.io.File;
+import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.FileReader;
@@ -327,6 +328,12 @@ public class JythonUtil {
                             throw new IOException("Unable to makeHomeFor "+ff5);
                         }
                     }
+                    if ( ff5.exists() ) {
+                        logger.fine("already have file, skip...");
+                        in.close();
+                        continue;
+                    }
+                    
                     try (FileOutputStream out = new FileOutputStream( ff5 ) ) {
                         transferStream(in,out);
                     } finally {
